Read package metadata from tar.bz2
