+++ pom.xml
@@ -92,6 +92,10 @@ SOFTWARE.
       <artifactId>javax.json-api</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.glassfish</groupId>
+      <artifactId>javax.json</artifactId>
+    </dependency>
+    <dependency>
       <groupId>commons-cli</groupId>
       <artifactId>commons-cli</artifactId>
       <version>1.4</version>
@@ -158,6 +162,12 @@ SOFTWARE.
       <artifactId>jsonassert</artifactId>
       <version>1.5.0</version>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-compress</artifactId>
+      <version>1.14</version>
+      <scope>compile</scope>
     </dependency>
   </dependencies>
   <build>
