--- src/test/com/artipie/conda/package-info.java
@@ -1,11 +1,0 @@
-/*
- * The MIT License (MIT) Copyright (c) 2020-2021 artipie.com
- * https://github.com/artipie/conda-adapter/LICENSE
- */
-
-/**
- * Conda adapter files tests.
- *
- * @since 0.1
- */
-package com.artipie.conda;
